Moved platform shutdown code from OrderProcessor to Customer
