+++ src/main/java/org/right_brothers/agents/BakeryCustomerAgent.java
@@ -1,10 +1,16 @@
 package org.right_brothers.agents;
 
 import jade.core.Agent;
+import jade.content.lang.Codec;
+import jade.content.lang.sl.SLCodec;
+import jade.content.onto.basic.Action;
 import jade.core.AID;
 import jade.core.behaviours.*;
 import jade.domain.FIPAAgentManagement.*;
+import jade.domain.JADEAgentManagement.JADEManagementOntology;
+import jade.domain.JADEAgentManagement.ShutdownPlatform;
 import jade.domain.FIPAException;
+import jade.domain.FIPANames;
 import jade.domain.DFService;
 import jade.lang.acl.ACLMessage;
 import jade.lang.acl.MessageTemplate;
@@ -13,11 +19,13 @@ import jade.lang.acl.MessageTemplate;
 @SuppressWarnings("serial")
 public class BakeryCustomerAgent extends Agent {
     private AID[] sellerAgents;
+    private static int totalAgents;
 
     protected void setup() {
         System.out.println("\tCustomer-agent "+getAID().getLocalName()+" is born.");
-
-        this.publishOrderProcessingAID();
+        totalAgents++;
+        
+        this.publishCustomerAID();
         try {
             Thread.sleep(1000);
         } catch (InterruptedException e) {
@@ -27,8 +35,8 @@ public class BakeryCustomerAgent extends Agent {
         Object[] args = getArguments();
         if (args != null && args.length > 0) {
             for (int i = 0; i < args.length; i++) {
-                String Bread_title = (String) args[i];
-                addBehaviour(new  RequestPerformer(Bread_title));
+                String order = (String) args[i];
+                addBehaviour(new  RequestPerformer(order));
             }
         }
         else {
@@ -47,7 +55,7 @@ public class BakeryCustomerAgent extends Agent {
         System.out.println("\t" + getAID().getLocalName() + ": Terminating.");
     }
 
-    protected void publishOrderProcessingAID(){
+    protected void publishCustomerAID(){
         DFAgentDescription dfd = new DFAgentDescription();
         dfd.setName(getAID());
         ServiceDescription sd = new ServiceDescription();
@@ -86,10 +94,10 @@ public class BakeryCustomerAgent extends Agent {
 	private class RequestPerformer extends Behaviour {
 		private MessageTemplate mt;
 		private int step = 0;
-        private String Bread_title;
+        private String order;
 
-        RequestPerformer (String Bread_title) {
-            this.Bread_title = Bread_title;
+        RequestPerformer (String order) {
+            this.order = order;
         }
 
 		public void action() {
@@ -99,7 +107,7 @@ public class BakeryCustomerAgent extends Agent {
 				for (int i = 0; i < sellerAgents.length; ++i) {
 					cfp.addReceiver(sellerAgents[i]);
 				}
-				cfp.setContent(this.Bread_title);
+				cfp.setContent(this.order);
 				cfp.setConversationId("Bread-trade");
 				cfp.setReplyWith("cfp"+System.currentTimeMillis()); // Unique value
 				myAgent.send(cfp);
@@ -112,6 +120,7 @@ public class BakeryCustomerAgent extends Agent {
 				if (reply != null) {
 					if (reply.getPerformative() == ACLMessage.CONFIRM) {
                         System.out.println("\t" + myAgent.getLocalName() + " received confirmation from " + reply.getSender().getLocalName());
+                        totalAgents--;
 						step = 2;
 					}
 				}
@@ -120,17 +129,33 @@ public class BakeryCustomerAgent extends Agent {
 				}
 				break;
             default:
-                step = 0;
                 break;
 			}
 		}
 		public boolean done() {
             if (step == 2) {
-                myAgent.doDelete();
+            	if(totalAgents == 0) {
+            		shutdown();
+            	}
                 return true;
             }
             return false;
 		}
+		
+		public void shutdown() {
+            ACLMessage shutdownMessage = new ACLMessage(ACLMessage.REQUEST);
+            Codec codec = new SLCodec();
+            myAgent.getContentManager().registerLanguage(codec);
+            myAgent.getContentManager().registerOntology(JADEManagementOntology.getInstance());
+            shutdownMessage.addReceiver(myAgent.getAMS());
+            shutdownMessage.setLanguage(FIPANames.ContentLanguage.FIPA_SL);
+            shutdownMessage.setOntology(JADEManagementOntology.getInstance().getName());
+            try {
+                myAgent.getContentManager().fillContent(shutdownMessage,new Action(myAgent.getAID(), new ShutdownPlatform()));
+                myAgent.send(shutdownMessage);
+            }
+            catch (Exception e) {}
+        }
 	} // End of inner class RequestPerformer
